requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-dotenv==1.0.1
openai==1.59.9
gspread==6.1.4
google-auth==2.37.0
python-multipart==0.0.20
